Ãrefactor: split core
