+++ examples/omni/data_classes/_core/helpers.py
@@ -1,0 +1,54 @@
+from __future__ import annotations
+
+from cognite.client import data_modeling as dm
+from .base import DomainModel, T_DomainModel
+from .constants import DEFAULT_INSTANCE_SPACE
+
+
+def as_node_id(value: dm.DirectRelationReference) -> dm.NodeId:
+    return dm.NodeId(space=value.space, external_id=value.external_id)
+
+
+def as_direct_relation_reference(
+    value: dm.DirectRelationReference | dm.NodeId | tuple[str, str] | None
+) -> dm.DirectRelationReference | None:
+    if value is None or isinstance(value, dm.DirectRelationReference):
+        return value
+    if isinstance(value, dm.NodeId):
+        return dm.DirectRelationReference(space=value.space, external_id=value.external_id)
+    if isinstance(value, tuple):
+        return dm.DirectRelationReference(space=value[0], external_id=value[1])
+    raise TypeError(f"Expected DirectRelationReference, NodeId or tuple, got {type(value)}")
+
+
+def as_pygen_node_id(value: DomainModel | dm.NodeId | str) -> dm.NodeId | str:
+    if isinstance(value, str):
+        return value
+    elif value.space == DEFAULT_INSTANCE_SPACE:
+        return value.external_id
+    elif isinstance(value, dm.NodeId):
+        return value
+    return value.as_id()
+
+
+def are_nodes_equal(node1: DomainModel | str | dm.NodeId, node2: DomainModel | str | dm.NodeId) -> bool:
+    if isinstance(node1, (str, dm.NodeId)):
+        node1_id = node1
+    else:
+        node1_id = node1.as_id() if node1.space != DEFAULT_INSTANCE_SPACE else node1.external_id
+    if isinstance(node2, (str, dm.NodeId)):
+        node2_id = node2
+    else:
+        node2_id = node2.as_id() if node2.space != DEFAULT_INSTANCE_SPACE else node2.external_id
+    return node1_id == node2_id
+
+
+def select_best_node(
+    node1: T_DomainModel | str | dm.NodeId, node2: T_DomainModel | str | dm.NodeId
+) -> T_DomainModel | str | dm.NodeId:
+    if isinstance(node1, DomainModel):
+        return node1  # type: ignore[return-value]
+    elif isinstance(node2, DomainModel):
+        return node2  # type: ignore[return-value]
+    else:
+        return node1
